refactor: moved suggestions to metatags handler instead of common
